move the test to .internal.js
