+++ packages/react-dom/src/__tests__/ReactTestUtils-test.js
@@ -14,7 +14,6 @@ let React;
 let ReactDOM;
 let ReactDOMServer;
 let ReactTestUtils;
-let ReactFeatureFlags;
 
 function getTestDocument(markup) {
   const doc = document.implementation.createHTMLDocument('');
@@ -29,8 +28,6 @@ function getTestDocument(markup) {
 
 describe('ReactTestUtils', () => {
   beforeEach(() => {
-    ReactFeatureFlags = require('shared/ReactFeatureFlags');
-    ReactFeatureFlags.enableHooks = true;
     createRenderer = require('react-test-renderer/shallow').createRenderer;
     React = require('react');
     ReactDOM = require('react-dom');
@@ -517,28 +514,5 @@ describe('ReactTestUtils', () => {
 
     ReactTestUtils.renderIntoDocument(<Component />);
     expect(mockArgs.length).toEqual(0);
-  });
-
-  it('flushEffects should run any enqueued effects', () => {
-    let effected = false;
-    let layoutEffected = false;
-    function App() {
-      React.useEffect(() => {
-        effected = true;
-      });
-      React.useLayoutEffect(() => {
-        layoutEffected = true;
-      });
-      return null;
-    }
-    ReactTestUtils.renderIntoDocument(<App />);
-    // effects haven't fired yet
-    expect(effected).toBe(false);
-    // layout effects have, however
-    expect(layoutEffected).toBe(true);
-
-    ReactTestUtils.flushEffects();
-    // effects have now fired
-    expect(effected).toBe(true);
   });
 });
